requirements.txt
boto3==1.34.34
aws-cdk-lib==2.124.0
constructs>=10.0.0
python-dotenv==1.0.1
requests==2.31.0 